1 Resume the renaming 2 Move the test case to sql/core
